Removed ancient and largely unused "started" message.
Moved "connection succesful" message to a point later in the program when we
actually know that it was succesful.  It was saying it regardless... bleah.

Fixed bugs in info where table and referee jids had been moved from the heap
to the new user object.

Changed calls to make_rpc_request use the new "handler" argument (instead of
storing a "response_id" for later use).

Fixed the roster display list to handle users without resources.

Made jabber_authed actually check to see if authentication was succesful.
It now reports sucess or failure and exits on failure.

Dumped the specialized "handle_rpc_response" in favor of new system.
Defined a bunch of default rpc_response's.

Removed extraneous defined check (apparently added to supress a warning;
it didn't work.)