What this does is allow you to have some hash values act like they are tied scalars without actually having to go through the trouble of making them really be tied scalars.
By default the tied hash works exactly like a normal hash.  Its behavior
changes when you use a value of an array with exactly two code blocks in it.
When you do this it uses the first as the get routine and the second as the
set routine.  Any future gets or sets to this key will be mediated via
these subroutines.