Changed the constructors of GameUI (and SVGCanvas) to take arguments for a message-printing service, and a token-translating service.
Added Javascript functions message(), localize(), and literalmessage().
(Which use the above.)

Added a Javascript-readable property info.seat; it currently returns
a dummy value.

Factored some code out of the TableWindow constructor. There is now
a utility function for "find the subdirectory of this dir that actually
contains stuff", and another for "find the entry in this directory which
case-insensitively matches a string." (Both are useful for searching
unpacked ZIP files.)

The "cannot create table" error dialog now has "Cannot Create Table"
as its window title (instead of the generic "Error".)