Added a "Join Table At..." menu option, to join an existing table (given the MUC JID).
As part of this, moved the gameTable.join() call back earlier in the
join process. This is safe because of the new scheme: the referee will
only send seating/state data after the client requests it. TableWindow
now sends this request (volity.send_state) after the UI is completely
built and the MUC is joined.

As part of *that*, added a ReadyListener interface to GameTable. A
caller can now listen for notification that the GameTable has joined
and located its referee.