Updated RPCRequester to catch XMPP stanza-level errors as well as RPC faults. Also made it generate XMPPException for timeout errors, instead of RPCException.
Fixed NewTableAtDialog to display a clean error for 404-JID-not-found.
It also displays reasonably clean errors for other error cases, including
timeouts.

Also added an assert to GameUI to catch a weird scripting bug which has
been fixed for a long time, but I accidentally re-triggered it while
doing other work, and it made me sad.