The SeatChart now colors player names. (The referee is always gray.)
UserColorMap is somewhat smarter now: if you give it a JID and a
nickname for a user, it maps the same color to both. The SeatChart
(which tracks real JIDs) makes use of this facility. As a result,
players at a table can change nickname without changing color.

Changed the algorithm UserColorMap uses to invent new colors. The hues
were good, but the brightness values were uneven. Now they're more even
around the spectrum.