The Smack work begins to pay off. We now parse for JEP-0115 tags in incoming packets; they appear as CapPacketExtensions.
Making use of this, the roster now displays a distinct icon for
buddies who are logged on with a Volity client.

The GameTable also uses this to recognize the referee without doing an
unnecessary disco check. (It falls back to the disco method if the
owner sends no JEP-0115 data. So this change won't break anything.)

There is a RosterPanel method listVolityClientResources(user), which
returns all the full JIDs of that user which are Volity clients. (The
invitation-sender will make use of this, although it doesn't yet.)