Considerably cleaned up the handling of failed new-table and join-table requests. We now always leave the MUC and close the table window in error cases.
(Big exception: if the join() of a new-table fails, we wind up with a hung
table window. I am not comfortable trying to fix this, because whenever
I test it, the conference.volity.net server dies. No can test, no can fix.
Will file that one separately.)

This takes care of bug 1305495 ("Joining a table with an existing
nickname kills the app".)