Added the content handler for MIME type "application/x-volity-command-stub".
This is less useful than you think. If a URL object (inside Javolin) has
that MIME type, then calling url.getContent() will return a CommandStub.
But nothing in Javolin creates or uses URL objects like that.

On OSX, it is possible to set up a Javolin app bundle so that double-clicking
a .volcom file hands it to Javolin. This doesn't have any practical value
either.