Fixed bug [ 1308404 ]: when you "join" a table, the status line doesn't update
I fixed this by increasing the timeout on disco queries from 5 seconds to
30. We may need more, for games that shove a lot of info into state bursts.
(We're sending out a query when we join a game, and it winds up waiting
on the state burst.)

I made that query happen in a background thread, but it didn't make the
waitingness go away. Still a good change to have made, though.